+++ rsconnect/api.py
@@ -5,6 +5,11 @@ from rsconnect.http_support import HTTPResponse, HTTPServer, append_to_path
 from rsconnect.log import logger
 from rsconnect.models import AppModes
 
+_error_map = {
+    4: 'This content has been deployed before but could not be found on the server.\nUse the --new option to '
+       'deploy it as new content.'
+}
+
 
 class RSConnectException(Exception):
     def __init__(self, message):
@@ -34,7 +39,12 @@ class RSConnectServer(object):
             # also catches all error conditions which we will report as "not running Connect".
             else:
                 if response.json_data and 'error' in response.json_data:
-                    raise RSConnectException('The Connect server reported an error: %s' % response.json_data['error'])
+                    code = response.json_data['code']
+                    if code in _error_map:
+                        error = _error_map[code]
+                    else:
+                        error = 'The Connect server reported an error: %s' % response.json_data['error']
+                    raise RSConnectException(error)
                 raise RSConnectException(
                     'Received and unexpected response from RStudio Connect: %s %s' % (response.status, response.reason)
                 )